+++ generate-xml/domain.nix
@@ -213,6 +213,7 @@ let
                   targetelem
                   (subelem "readonly" [ ] [ ])
                   addresselem
+                  (subelem "boot" [ (subattr "order" typeInt) ][])
                 ]
               )
               (subelem "filesystem" [ (subattr "type" typeString) (subattr "accessmode" typeString) ]
@@ -252,6 +253,7 @@ let
                 ]
                 [
                   (subelem "source" [] [ addresselem ])
+                  (subelem "boot" [ (subattr "order" typeInt) ][])
                 ]
               )
               (subelem "shmem"