+++ generate-xml/domain.nix
@@ -16,6 +16,29 @@ let
         (subelem "currentMemory" [ (subattr "unit" typeString) ] (sub "count" typeInt))
         (subelem "vcpu" [ (subattr "placement" typeString) ] (sub "count" typeInt))
         (subelem "iothreads" [ ] (sub "count" typeInt))
+        (subelem "cputune" [ ] [
+          (subelem "vcpupin" [ (subattr "vcpu" typeInt) (subattr "cpuset" typeString) ] [ ])
+          (subelem "emulatorpin" [ (subattr "cpuset" typeString) ] [ ])
+          (subelem "iothreadpin" [ (subattr "iothread" typeInt) (subattr "cpuset" typeString) ] [ ])
+          (subelem "shares" [ ] typeInt)
+          (subelem "period" [ ] typeInt)
+          (subelem "quota" [ ] typeInt)
+          (subelem "global_period" [ ] typeInt)
+          (subelem "global_quota" [ ] typeInt)
+          (subelem "emulator_period" [ ] typeInt)
+          (subelem "emulator_quota" [ ] typeInt)
+          (subelem "iothread_period" [ ] typeInt)
+          (subelem "iothread_quota" [ ] typeInt)
+          (subelem "vcpusched" [ (subattr "vcpus" typeString) (subattr "scheduler" typeString) (subattr "priority" typeInt) ] [ ])
+          (subelem "iothreadsched" [ (subattr "iothreads" typeInt) (subattr "scheduler" typeString) ] [ ])
+          (subelem "cachetune" [ (subattr "vcpus" typeString) ] [
+            (subelem "cache" [ (subattr "id" typeInt) (subattr "level" typeInt) (subattr "type" typeString) (subattr "size" typeInt) (subattr "unit" typeString) ] [ ])
+            (subelem "monitor" [ (subattr "level" typeInt) (subattr "vcpus" typeString) ] [ ])
+          ])
+          (subelem "memorytune" [ (subattr "vcpus" typeString) ] [
+            (subelem "node" [ (subattr "id" typeInt) (subattr "bandwidth" typeInt) ] [ ])
+          ])
+        ])
 
         (subelem "sysinfo" [ (subattr "type" typeString) ] [
           (subelem "bios" [] [