+++ CHANGELOG.md
@@ -7,6 +7,7 @@
         * Allow in cpu section, customization of the `feature` and `cache` option
         * Add section `sysinfo`
         * add bootorder for device `disk` and `hostdev`
+        * add section `iothreads`
         * Updated section `memoryBacking`
 
 ## [0.5.0]