+++ CHANGELOG.md
@@ -8,6 +8,7 @@
         * Add section `sysinfo`
         * add bootorder for device `disk` and `hostdev`
         * add section `iothreads`
+        * Updated section `memoryBacking`
 
 ## [0.5.0]
 * Lib: