+++ CHANGELOG.md
@@ -9,6 +9,7 @@
         * add bootorder for device `disk` and `hostdev`
         * add section `iothreads`
         * Updated section `memoryBacking`
+        * Update section `clock`
         * Allow in feature section, customization of the `ioapic` option
         * add section `cputune`