+++ CHANGELOG.md
@@ -12,6 +12,8 @@
         * Update section `clock`
         * Allow in feature section, customization of the `ioapic` option
         * add section `cputune`
+        * add section `qemu:override`
+        * Updated section `hostdev`
 
 ## [0.5.0]
 * Lib: